Split streaming behaviour tests into focused modules
The streaming pull-parser integration tests had grown to 869 lines in a
single file, breaching the 400-line-per-file maintainability rule in
AGENTS.md. Adding the namespaced parser coverage (#49) pushed it further
over, so reorganise the suite before it accretes more cases.

The BDD scenarios must share one test binary because rstest-bdd resolves
steps through a global registry, so the harness is split across module
files rather than separate binaries:

  - streaming/support.rs: captured state, named XML fixtures, and the
    rstest fixtures that seed each scenario.
  - streaming/steps.rs: the shared given/when/then step definitions.
  - streaming/namespace.rs: the namespace-aware assertion step and
    scenario, isolated so future namespace cases land here instead of
    enlarging the behaviour file.
  - streaming_behaviour.rs: the binary root binding the core scenarios.

The harness-free structural tests (divisions, nested divisions, entity
references) drive the public parser API directly, so they move to their
own top-level binary streaming_structure.rs, matching the existing
streaming_empty_elements.rs convention.

Behaviour is unchanged: the full suite still runs 456 tests, and every
file is now within the 400-line limit.
